feat(MiniApp(Token)): Make token mini-app
