+++ src/interfaces/http/router.js
@@ -46,7 +46,7 @@ module.exports = ({ config, logger, database }) => {
    */
 
   apiRouter.use('/', controller('index'))
-  apiRouter.use('/token', controller('token'))
+  apiRouter.use('/token', controller('token').router)
   apiRouter.use('/users', controller('user').router)
   apiRouter.use('/companies', controller('company').router)
